requirements.txt
annotated-types==0.7.0
bitarray==3.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
crcmod==1.7
greaseweazle @ git+https://github.com/keirf/greaseweazle@dcb3d5d9d71b58d51aa22938cf96272a3b507292
idna==3.10
pydantic==2.11.3
pydantic_core==2.33.1
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
pyserial==3.5
PyYAML==6.0.2
requests==2.32.3
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
